requirements.txt
django>=2.0
mysqlclient==1.3.13
mysql-connector-python==8.0.13
Pillow==5.3.0
easy-thumbnails==2.5
django-ckeditor>=5.3.0
django-filebrowser-no-grappelli